fix(tearsheet): added utils.ts
